backend/auth/requirements.txt
aiofiles==24.1.0
aiosmtplib==5.0.0
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==5.0.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.118.3
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
iniconfig==2.3.0
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.3
oauthlib==3.3.1
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
